+++ package.json
@@ -9,6 +9,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "react-speech-recognition": "^3.10.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
