+++ src/Speech-Recognization/Speech.css
@@ -1,0 +1,11 @@
+.container1{
+    margin: auto;
+    padding: auto;
+    height: auto;
+    width: 350px;
+}
+.Main-Context{
+    width: 300px;
+    height: 250px;
+    border: 3px solid green;
+  }
