+++ src/component/App.css
@@ -3,6 +3,7 @@
   float: left;      
 }
 
+
 @keyframes App-logo-spin {
   from {
     transform: rotate(0deg);
